Added purse functions, started belongings sheet
- Fixed a few issues
- Renamed Inventory to Equipment
- Renamed Items to Belongings
- ItemInstance.isParryingWeapon integration not complete
- Reworked equipment list layout

+++ src/utils/ItemUtils.ts
@@ -23,15 +23,15 @@ export const convertToSave = (item: ItemEditorInstance) => ({
 	damageDiceNumber: item.damageDiceNumber ? Number.parseInt(item.damageDiceNumber.replace(/\,/, '.')) : 0,
 	damageFlat: item.damageFlat ? Number.parseInt(item.damageFlat.replace(/\,/, '.')) : 0,
 	enc: item.enc ? Number.parseInt(item.enc.replace(/\,/, '.')) : 0,
-	length: item.length ? Number.parseInt(item.length.replace(/\,/, '.')) : 0,
+	length: item.length ? Number.parseFloat(item.length.replace(/\,/, '.')) : 0,
 	amount: item.amount ? Number.parseInt(item.amount.replace(/\,/, '.')) : 0,
 	pa: item.pa ? Number.parseInt(item.pa.replace(/\,/, '.')) : 0,
-	price: item.price ? Number.parseInt(item.price.replace(/\,/, '.')) : 0,
+	price: item.price ? Number.parseFloat(item.price.replace(/\,/, '.')) : 0,
 	pro: item.pro ? Number.parseInt(item.pro.replace(/\,/, '.')) : 0,
 	range: item.range.map(e => e ? Number.parseInt(e.replace(/\,/, '.')) : 0),
 	reloadTime: item.reloadTime ? Number.parseInt(item.reloadTime.replace(/\,/, '.')) : 0,
 	stp: item.stp ? Number.parseInt(item.stp.replace(/\,/, '.')) : 0,
-	weight: item.weight ? Number.parseInt(item.weight.replace(/\,/, '.')) : 0
+	weight: item.weight ? Number.parseFloat(item.weight.replace(/\,/, '.')) : 0
 });
 
 export const containsNaN = (item: ItemInstance) => {
